Move functions to local utils file
